Experiment with running the main fetch-execute-decode for litmus
The _TLB magic 'register' is problematic, because it's huge and written
to at least 1024 times which turns into probably billions of times when
combined with the current non-determinism in TakeReset, that quickly
exhausts all available memory. Not putting _TLB in the trace fixes the
memory usage, but it's still unlikely to get through TakeReset until
all the the non-determinism is gone.

Also remove some experiments that aren't likely to be used