Add some more debugging features
Rather than generate a full trace which is not useful for more than
small examples, allow specifying individual functions we want to gather
information for. By analysing the trace at those functions we can
determine if any symbolic arguments are influenced by uninitialized
registers or memory, which helps when trying to figure out the cause of
any non-determinism.