Generate linker script for relocations when assembling litmuses
When assembling a litmus test we need to make sure that branches have
the correct addresses, particularly branches between the code for
multiple threads, which does occur in some tests. To do this we generate
a simple linker script and use it to apply any relocations based upon
the offsets where each thread will be loaded into the model before
symbolic execution.