Work on web interface server
Can now send a architecture/litmus/cat triple to the server, which can
spawn workers to handle each request. The workers are each separate
processes restricted to use a fixed amount of CPU and memory using
setrlimit. Separate processes is so any bugs and crashes won't take
down the whole server. There is a fixed maximum number of workers, and
above that many concurrent requests any pending requests will wait
until the number of workers drops below the maximum. This should bound
the resource usage on the server.