Get more litmus tests working
To avoid becoming blocked on it, avoid dealing with the general case
of recursive herd relations and allow defining reflexive and/or
transitive relations at the toplevel, i.e.

transitive obs | dob | aob | bob as ob

rather than

let rec ob = obs
  | dob
  | aob
  | bob
  | ob; ob

Any such cat formula can in theory be represented in SMT, but the
complexity in the worst case appears quite bad. We can't directly
compute such a relation, although it is guaranteed the fixed point is
reached after 2^(n^2) iterations where n is the number of events
(there may be a better lower bound).