Change last_write_to_32 to (hopefully) fix timeouts on Z3 4.8.8
Some litmus tests were working on Z3 4.8.4 but timing out on Z3
4.8.8. Changing the definition of last_write_to_32 (predicate for the last
32-bit write to a location) seems to fix this for those
examples. Whereas previously we had a bits(64) -> bits(32) function
which had to be found as part of the model, we now have a
bits(64)*bits(32) -> bool predicate which is expanded wherever
used. This should be a bit more SMT friendly.