Some groundwork for wider concrete bitvectors (capabilities)
The .bits() function is gone from the BV trait, as it was left-over
from when there was just a concrete 64-bit specific type. In it's place
there is now a .lower_u64() method trait - which makes it much more
obvious what would happen with wider bitvectors. All uses of the .bits()
function have been checked in isla-lib and updated depending on whether
the lower_u64 semantics was really intended and changed if not.

The memory subsystem still assumes physical addresses are not more than
64-bits.

The length field of B64 has been changed to just 'len'. This is for
consistency with the rust standard library which uses this abbreviated
form.