Add version number computation in build.rs
The version number is stored in the ISLA_VERSION evironnement variable.
It first try `git decribe --dirty` and if git is unavaible (when
installing from crates.io) fallback on CARGO_PKG_VERSION)

isla-client now use that when answering the version command.

version starting with "dev-" will no longer appear after this commit and
are to be classified before versions starting with "v".