With a bit of hackery can now run some page table tests like CoWT
We don't necessarily get the right results - we can read
and write to the page tables, and observe behavior that is consistent
with page table reads and writes being the same as any other reads and
writes.

Running regular tests with address translation requires a bit less
trickery but is still not completely automatic.