Add a new registers.reset field to the config
Rather than registers.defaults which sets register values before any
symbolic execution, registers.resets defines a builtin `reset_registers`
which can be called by Sail code. Unlike registers.defaults this allows
for register fields to be set individually, i.e.

isla-footprint -R "PSTATE.EL = 0b10"

This is useful particularly for ARM as it allows registers to be set
after TakeReset is called. However we still want registers.defaults to
set certain registers like RVBAR which are used within TakeReset. It
also means we don't have to modify TakeReset if we want to run tests
with the MMU on and so on.