Stage 2 page tables support
Currently --armv8-page-tables options setups up stage 1 and stage 2
tables, mapping code and data into them as needed. They are set up so
the level 3 page table entries in the second stage can be invalidated
by writes, but this will need to be made more flexible.

Doing footprint analysis with address translation turned on doesn't work
well, so we have a separate configuration for that with the
--footprint-config option. This is a bit of a hack, as we have to load
the architecture twice and pass far too many things to some functions to
make this work.