Add Sail source locations to primitives and function calls
Originally each location contained the filename as a string, this seems
unwieldy when we are putting locations in far more places. Therefore
rather than

file.sail x:y - z:w

locations are now

n x:y-z:w

where n (along with x, y, z, and w) is a (natural) number. This number
is an index into the list of Sail files making up the IR which are
included via the toplevel `files` definition in the IR, which in the
above example would look something like:

files "other_file.sail" "file.sail"

in which case n would be 1

The list of files is stored in the Symbol table after processing the IR