Syntax for setting custom page table attributes
Also allow different attributes at stage 1 and stage 2 (previously they
could just be default and default or code and code).

Syntax is:

x |-> y with [field = value]

where field are the names used in the ARM ARM for each attribute. The
value must be a bitvector of the exact length of that field.

If the |-> operator would create a mapping in both S1 and S2, then this
will try to use the same attributes for both (which will usually cause an
error), to solve this you can do:

va |-> pa with [field = value] and default

which will use the first attributes at S1 and the second at S2.

In attributes like [AP = 0b00] all the other attributes not specified
will be the same as default.