Support for VMSA tests in web interface
Allow VMSA tests and options to be passed via the web interfaces

Uses new graph drawing code for web interface. Graphviz dot source is
generated server side. To allow the web interface to toggle the
relations on and off, we add a control_delimit option to the GraphOpts
struct that inserts special control characters to the graph output that
delimit which bits of the graphviz file correspond to which relation.