Repeatedly simplify the graph during partial linearization
Handle assertions correctly during partial linearization optimization.
This is done by converting assertions into control flow (i.e. a jump
instruction to an assertion failure statement), which means that the
control flow inherent in assertion failures becomes an explicit part of
the control flow graph and can be seen by the optimization.

Repeated graph simplification requires several functions to be
generalized as it allows multiple edges to the same node to share the
same 'edge number' which corresponds to the phi function arguments when
making the SSA graph executable. However this does allow it to be
usefully applied to larger examples in sail-arm