Use task fractions to decide when executor should finish
Previously there was some complex logic to try to figure out when every thread was finished, by
detecting when all worker threads were idle, had tried to steal work, and were still idle.
Now we have a fraction representing what percentage of each task has been worked on (and therefore
finished) this logic can be replaced by much simpler logic that just checks if all the tasks are done.

This also allows the state-merging after each instruction to work, as it's now possible to inject
a new task for the merged state at runtime which would previously have confused the above logic.

Note we use a bigint library for the fraction numerators (we store the denominator as an exponent 2 ^ N),
as if we split more than 64 times on a single branch then we wouldn't be able to represent the
fraction correctly. This came up in the test suite, and while a workaround can be implemented in
that case, it's probably better to just ensure it can't ever happen.